+++ changelog.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## 0.0.7
 - Feature: Updated taglib support for Liferay 7.1 and 7.2
+- Feature: Inspections for wrongly nested Liferay tags inside JSPs
 
 ## 0.0.6
 - Feature: File References for portlet.properties files